+++ scm-ui/ui-types/src/Repositories.ts
@@ -35,6 +35,10 @@ export type Repository = {
   _links: Links;
 };
 
+export type RepositoryCreation = Repository & {
+  contextEntries: { [key: string]: any };
+};
+
 export type RepositoryCollection = PagedCollection & {
   _embedded: {
     repositories: Repository[] | string[];